// src/components/LoadingCircle/enums.ts
export enum ELoadingCircleDefaults {
  SIZE = 64,
  STROKE_WIDTH = 6,
  DURATION_MS = 1200,
}

export enum ELoadingCircleAnimation {
  START_VALUE = 0,
  END_VALUE = 1,
}

export enum ELoadingCircleRotation {
  START = '0deg',
  END = '360deg',
}

export enum ELoadingCircleMath {
  DIVISOR_HALF = 2,
  MIN_RADIUS = 0,
  GAP_RATIO = 0.25,
}